Update README.md, pip_names => pip_and_srcs_lists
`pip_and_srcs_lists` expands upon the original `pip_names` rule to emit
a list of transitive source files local to the repository. This
completes the example to match the original question from Frank P.

Updates the `README.md` with this information, along with example
commands and output, and updates to the "history" section.